// src/app/models/HttpRequests/UpdateCompanyItem.ts
export class UpdateCompanyItem {
    userID: number;
    spesificitemID: number;
    itemID: number;
    item: string;
    description: string;
    tariff: number;
    type: string;
    usage: string;
    mIDP: string;
    pI: string;
    vulnerable: string;
    n521: string;
    n536: string;
    n31761: string;
    n31762: string;
    n31702: string;
}